Add a url-based SEG test pulling from ipfs
Work in progress: do not merge without further testing.

This example uses an ipfs-based url to pull the seg data
which does not appear to be completely reliable yet.
If we determine the ipfs approach is not worth pursuing, we can
still use the cliwrap-based testing approach to pull files from
other urls.

The cliwrap.py utility is copied into the dcmqi-bin directory
because it could prove useful for end-users who want to pull
data from urls as input to these programs.  Also the cliwrap approach
may be extended to handle file sets and directories in addition to
being able to upload results.