BUG: do not consider computed slice thickness for SEG
As discussed in #89, in the general case, we cannot rely on the slice thickness
computed from the positions of the individual frames, since frames may not be
contiguous. When source image references are available, and the source images
are also available, additional checks could be possible, but not in dcmqi.