ENH: add json compare scritp
This one is from https://github.com/ChannelIQ/jsoncompare with minor
modification submitted in this PR https://github.com/ChannelIQ/jsoncompare/pull/8.

Without the modification, it is required that the compared JSONs have identical
number of keys, which is not logical, since we are specifically interested in
ignoring specified keys.

Related to #97