BUG: cmake: Ensure git and svn executable are propagated
This problem was not revealed when Slicer extensions are
configured and built automatically by the Extensions build system setup
on the factory machines. In that case, variables like git and svn paths
are explicitly passed to the extension and added to the cache
using the "-D" cmake argument.

Instead, this commit fixes a problem occurring only when:

(1) svn or git is NOT in  the path

and

(2) extension is configured by only passing
    "-DSlicer_DIR:PATH=/path/to/Slicer-SuperBuild-Release/Slicer-build/"

In that case, because the values of Subversion_SVN_EXECUTABLE and
GIT_EXECUTABLE are retrieved in ExternalProjectDependency.cmake
in way that trigger https://gitlab.kitware.com/cmake/cmake/issues/15448

This commit explicitly pass then in SuperBuild.cmake despite the fact
they are already "mark_as_superbuild" in SlicerEnableExtensionTesting.cmake
used within SlicerConfig.cmake