ENH: update diffusion and perfusion parametric map contexts
These are derived automatically from the XML of the standard and/or correction
proposals that establish explicit connection between the quantities and quantity
units. CP-1665 describes the modifications to the standard related to the diffuson
context. CP with the modifications to the perfusion quantities is not yet assigned
a number.

Note from @dclunie:

> The choice to use an abbreviation (Ktrans, T1) or a fully spelled
> out Code Meaning ("Regional Cerebral Blood Flow") is historical
> and arbitrary, except in that "RCBF" would be very confusing
> (relative versus regional question).

This means that we might consider including definition of the quantity in the
context, since code meaning along can be too confusing.