ENH: include laterality only for the Anatomic Structure category
On 2/23/17 10:44 PM, Andrey Fedorov wrote:
>> The "paired" attribute controls whether or not the laterality
>> modifiers are added to the JSON file (i.e., only if "Y"),
>> and it doesn't really need to be in the JSON file itself,
>> except that it seemed reasonable to add it ... for that
>> matter, I was not sure which of the existing attributes your
>> code actually used.
>>
>
> This makes sense, except it is still not clear to me why there are
> essentially 4 options: paired attribute can be either present or not
> in JSON (e.g., it is absent for "Capillary"), and then if it is
> present, it can take one of the three values. What is the meaning of
> "paired" not present? Is it the same as "U" or something else?

"paired" no present means that the automated tool did not find it
in FMA and make a pairedness decision, and I have not yet added
my own manual determination to drive the JSON generation.

In the interim, it should be interpreted as U.

> Few clarification questions, beyond the one about pairedness:
>
>  - line 507: you specify pairedness for "Tissue" - is this right?

The way this works is that I search the ontology (FMA) for nodes
that have no mention of left and right but have left and right
children, set that node to "paired" and then propagate that back
up the class hierarchy.

This works well most of the time, but somewhere there must be
a distant descendant of "Tissue" (FMAID 9637) that has a left
or a right (or both) in it ... it will take a bit of digging
to find out what it is and see if it is a "mistake" in the
ontology or a flaw in my strategy (and/or the parts of the
FMA class tree I am allowing my code to search).

>  - line 616: pairedness for "Cartilage"?

That one is easy, it has children that are side-specific (e.g.,
the "fibrocartilage of left femur").

That doesn't make it useful to us though, and probably highlights
that unless the code is being used in the "Anatomical Structure"
category, it is neither useful or appropriate to populate the
laterality.

I have attached an updated JSON file that only produces laterality
for the "Anatomical Structure" category.

>  - "Body fluid" is also paired

Has children like "CSF in left side of third ventricle", etc.