cmake: Set package name including "version information"
The package has now the following format:

  dcmqi-vX.Y.Z-(linux|mac|win64)[<version-qualifier>].(tgz|zip)

where

  X.Y.Z corresponds to the major/minor/patch version define in the
  CMakeLists.txt

  <version-qualifier> is set based on rule described below.

This commit adds the target 'dcmqiConfigureVersion' that will determine
at build time if a *release* or a *development* package is being built. The
target is always built.

If a release is being built, the name of the package is:

  dcmqi-vX.Y.Z-(linux|mac|win64).(tgz|zip)

If a development package is being built, the appended <version-qualifier>
will be:

  "-<YYYYMMDD>-<sha_short>"

where

  <YYYYMMDD> is the date of the HEAD (last commit of the current Git branch)

  <sha_short> the short version of the SHA associated with HEAD

For example:

  dcmqi-1.0.5-linux-20170408-f06b084.tar.gz