ENH: make directory paths for SR converter optional
When directory path is not defined, use file names from the metadata file as is.
This will allow using full paths in the JSON directly, which can be important sometime.

As reported by @nolden and @JasminMetzger, sometime there could be files that have
identical names, or are located in different directories, that need to be referenced.