ENH: add support for initializing measurement properties
This aims to address #305. Schema extended to support numeric and text properties
of the measurement, as suggested by @dclunie in

https://github.com/QIICR/dcmqi/issues/305#issuecomment-335172718

Since these attributes are not supported by the dcmsr wrapper class, they are added
in a separate iteration by iterating over all measurement items using gotoAnnotatedNode(),
and setting measurement properties, when applicable.

Reading SR and saving JSON is not supported yet.