BUG: Fix dockcross script creation container cleanup
When the container is created, it does not have the name
'dockcross/manylinux-x64'. This is the name of the image as opposed to
the container. Docker gives a random, human-readable name
to a container if the `--name` flag is not provided.

Use the `--rm` flag so the container does not need to be removed.