ENH: revise the procedure for initializing tracking info
If neither TrackingID nor TrackingUID are initialized, then do nothing.

If TrackingID is initialized, generate TrackingUID and print a warning
alerting user that they should pay attention in case they encode longitudinal
data.

If TrackingUID is initialized, but TrackingID is not, reuse SegmentLabel as
TrackingID, and alert the user about that.