ENH: check both SCT and SRT codes
Sub-components of the SR document are recognized based on the codes for the
concepts describing the conent. With this commit, both SCT and SRT codes will
be checked against to accommodate both SRT-based legacy and modern SCT-based
SR documents (thus this is related to #386, but addresses it only for the
tid1500reader component of dcmqi).