WIP: Introduce OverlapUtil
Introducing a new class OverlapUtil that is able to check whether
segments of a given DICOM segmentation are overlapping.

It computes an overlap matrix denoting which segments overlap with each
other. Also, from that it can derive groups of segments that do not
overlap, e.g. for exporting them later into NRRD files or similar
formats that cannot deal with overlapping segments.