Use Label IDs as Segment Numbers.
When converting from ITK to DICOM, allow to use Label IDs as Segment
Numbers so that even if reading from various NRRD inputs, the Segments
in the DICOM file will still use the Label IDs from the NRRD files.

This is particularly helpful for testing since it allows a roundtrip test
from DICOM -> multiple NRRD files using potentially multiple segments
each -> back to DICOM. If no mapping to the Label IDs takes place, the
converter will use ascending Segment Numbers, so the order will depend
on the read order, which again usually depends on the order in the JSON
meta information.

This sorting behavior is disabled by default. It is available as
cli option (--sortByLabelID) which is handed to the library through the
Itk2DicomConverter::itkimage2dcmSegmentation() call.