Fix order of pixel spacing converting ITK->DICOM.
The order in which the values for pixel spacing in ITK and DICOM are
stored is different. In DICOM the first value is the row spacing
(i.e. Y axis) while in ITK the first value is spacing along the X axis.

Reading asymmetric spacings from DICOM segmentations has already been
fixed in 46c6f0.

This commit fixes writing DICOM segmentation objects with asymmetric
spacings.