Switch to official DCMTK 3.6.8++ (June 6th 2025).
The selected intermediate commit (30cd13) was mainly selected since it
includes an earlier commit for the C++ build that replaces the need for
the former patched DCMTK version from the DCMTK repo and a few commits
more to fix shared library build.

The updated DCMTK also allows to import dataset references from
DcmItem instead of DcmDataset for the Derivation Image Functional
Group which is easier to use when creating references from DICOM
data memory.