DICOM binary segmentation to labelmap converter.
Introduced new command line tool bin2labelsegimage that converts DICOM
binary segmentation objects  (Segmentation Storage SOP Class) to labelmap
(Labelmap Segmentation Storage SOP Class) objects.

General usage is:

bin2labelsegimage --inputDICOM binary_seg.dcm
--outputDICOM labelmap_seg.dcm

It will then try to convert the given binary segmentation to a labelmap
segmentation.

By default it will use MONOCHROME2 as output color model. For using
palette color instead, option  --usePalette is available. For palette, an
ICC profile is required by the standard. For now, the behavior is to use
colors as specified in the binary segmentation (though i think its not
heavily tested) if all segments have assigned a recommended color in the
input, and otherwise write a default ICC profile and use random colors per
segment.

The converter automatically decides whether it produces 16 bit or 8 bit
labelmaps.

It also adds a Derivation Image entry in the Derivation Image Functional
Group and adds the original binary segmentation to the list of inputs.
Right now, this is only done if the Derivation Image FG in the input is
shared or not present. In case its already per-frame, no addition is done,
which can be improved in the future.

The new labelmap object is placed into a new series but stays in the same
study as the input segmentation.